build: Install local editable after installing requirements.
I would add the local project to dev.in but that results in the compiled
output including a fully qualified path to my local version of the
project right now which is not helpful.

Adding `-e .` to `dev.in` puts a `-e file://my/local/path/to/sample-plugin`
in the dev.txt file which is not helpful.