feat: Add links to course home and about.
Add a menu linking to the course about page. This page is actually being
overridden by our plugin so we can now easily show on one page most of
the overrides that our plugin makes.

We also added some links to the actual course content given that this is
the learner home and we should probably let them access the course
content.