build: Move dynamic settings to pyproject.toml
Moving the necessary dynamic settings from setup.py to pyproject.toml

A few settings were not migrated because their values are the default
when a pyproject.toml file is used and a few were not moved because they
were deprecated.

Full list that I used for reference is here: https://setuptools.pypa.io/en/latest/userguide/pyproject_config.html#setuptools-specific-configuration

The constraint.txt reference is not PEP508 compliant so is not supported
for the files that are read in by setuptools.  So I moved the constraint
into the makefile call for base.in so it could still serve as the source
of truth for dependencies for this project.